Add Symbol class for template declarations.
Currently un-used. In first iteration the templates can not be
overloaded. Simplified algorithm would sound as follows. If at call site
function name "foo" is found to be template name then:
1. Unify types of concrete and formal arguments and get subsitution if
   that succeeds.
2. If "foo" has not been instantiated with given substitution then
   instantiate the template.
3. Generate code for the call.
After code for currently active procedure has been generated mark that
done and continue with next procedure in the work list.

There are few questions that need answering before I can get to some
serious coding.  Do I need to copy the AST to make instantiation easier,
or is it sufficient to give type checker the substitionion? Must I get
rid of AST rewriting (rewriting seems to be fine if we copy the AST, but
for some reason I'd like to avoid doing that)?