[scc] Refactored VMInstruction and VMValue classes
A VMInstruction basically just converts a list of operands to strings and
outputs them. Since the operands are not used in any other way it does not make
much sense to hold (fragile) pointers to such specific classes where one just
needs to store pointer to streamable things. Using plain strings would not work
because of virtual registers which are initialized after the resulting operands
are pushed to VMInstruction.

This commit creates an OStreamable superclass for such objects. The OStreamable
objects are created by the respective VMValue subclass using shared pointers
and later passed used by the VMInstruction object.